+++ django_project/frontend/src/pages/Admin/Indicator/List/index.jsx
@@ -3,29 +3,18 @@ import React from 'react';
 import { render } from '../../../../app';
 import { store } from '../../../../store/admin';
 import { pageNames } from '../../index';
-import AdminList from "../../Components/List";
+import AdminList, { COLUMNS } from "../../Components/List";
 
 import './style.scss';
-
-
-const columns = [
-  { field: 'id', headerName: 'id', hide: true, width: 30 },
-  { field: 'name', headerName: 'Indicator Name', flex: 1 },
-  { field: 'description', headerName: 'Description', flex: 1 },
-  { field: 'group', headerName: 'Category', flex: 1 }
-]
 
 /**
  * Indicator List App
  */
 export default function IndicatorList() {
   return <AdminList
-    columns={columns}
+    columns={COLUMNS(urls.admin.indicatorList)}
     pageName={pageNames.Indicators}
     listUrl={urls.api.list}
-    editUrl={urls.api.edit}
-    detailUrl={urls.api.detail}
-    redirectUrl={urls.admin.indicatorList}
   />
 }
